Nuget

+++ Services/Controllers/ServicesController.cs
@@ -1,7 +1,12 @@
 using AutoMapper;
+using MassTransit;
+using MessagingModels;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services.Data;
 using Services.Dtos;
+using Services.Extensions;
 using Services.Models;
 
 namespace Services.Controllers
@@ -15,19 +20,22 @@ namespace Services.Controllers
 		private readonly IPhotoRepo _photoRepository;
 		private readonly IPhotoService _photoService;
 		private readonly IMapper _mapper;
+		private readonly IPublishEndpoint _publishEndPoint;
 
 		public ServicesController(
 			IServiceRepo serviceRepository,
 			IServiceCategoryRepo serviceCategoryRepository,
 			IPhotoRepo photoRepository,
 			IPhotoService photoService,
-			IMapper mapper)
+			IMapper mapper,
+			IPublishEndpoint publishEndPoint)
 		{
 			_serviceRepository = serviceRepository;
 			_serviceCategoryRepository = serviceCategoryRepository;
 			_photoRepository = photoRepository;
 			_photoService = photoService;
 			_mapper = mapper;
+			_publishEndPoint = publishEndPoint;
 		}
 
 		/// <summary>
@@ -37,10 +45,9 @@ namespace Services.Controllers
 		/// <param name="serviceCreateDto"></param>
 		/// <returns></returns>
 		[HttpPost]
-		public ActionResult CreateService(int userId, ServiceCreateDto serviceCreateDto)
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+		public async Task<ActionResult> CreateService(int userId, ServiceCreateDto serviceCreateDto)
 		{
-			// TODO: validate user is logged in
-
 			// check if category exists
 			var serviceCategory = _serviceCategoryRepository.GetServiceCategory(serviceCreateDto.ServiceCategoryId);
 
@@ -49,16 +56,37 @@ namespace Services.Controllers
 				return BadRequest("Service category does not exist");
 			}
 
+			// check if userId is the same as the current user id
+			var userIdInToken = this.User.GetId();
+
+			if (userIdInToken != userId.ToString())
+			{
+				return Unauthorized();
+			}
+
 			var serviceModel = _mapper.Map<Service>(serviceCreateDto);
 
 			_serviceRepository.CreateService(userId, serviceModel);
 			_serviceRepository.SaveChanges();
 
+			// Publish ServiceCreated event
+			await _publishEndPoint.Publish<ServiceCreated>(new
+			{
+				Id = serviceModel.Id,
+				Info = serviceModel.Info,
+				UserId = serviceModel.UserId,
+				ServiceCategoryId = serviceModel.ServiceCategoryId
+			});
+
 			var serviceReadDto = _mapper.Map<ServiceReadDto>(serviceModel);
 
 			return CreatedAtRoute(
 				nameof(GetService),
-				new { userId = userId, serviceId = serviceReadDto.Id },
+				new
+				{
+					userId = userId,
+					serviceId = serviceReadDto.Id
+				},
 				serviceReadDto
 			);
 		}
@@ -68,10 +96,9 @@ namespace Services.Controllers
 		/// </summary>
 		/// <returns></returns>
 		[HttpGet]
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
 		public ActionResult<IEnumerable<ServiceReadDto>> GetServices(int userId)
 		{
-			// TODO: validate user is logged in
-
 			var services = _mapper.Map<IEnumerable<ServiceReadDto>>(_serviceRepository.GetServicesForUser(userId));
 
 			return Ok(services);
@@ -84,10 +111,9 @@ namespace Services.Controllers
 		/// <param name="serviceId"></param>
 		/// <returns></returns>
 		[HttpGet("{serviceId}", Name = "GetService")]
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
 		public ActionResult<IEnumerable<ServiceReadDto>> GetService(int userId, int serviceId)
 		{
-			// TODO: validate user is logged in
-
 			// get service
 			var service = _serviceRepository.GetService(serviceId, userId);
 
@@ -107,15 +133,27 @@ namespace Services.Controllers
 		/// <param name="serviceUpdateDto"></param>
 		/// <returns></returns>
 		[HttpPut("{serviceId}")]
-		public ActionResult<ServiceReadDto> UpdateService(int userId, int serviceId, ServiceUpdateDto serviceUpdateDto)
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+		public async Task<ActionResult<ServiceReadDto>> UpdateService(int userId, int serviceId, ServiceUpdateDto serviceUpdateDto)
 		{
-			// TODO: validate user is logged in (JWT)
-			// TODO: validate user is owner (if not unauthorized)
+			// check if userId is the same as the current user id
+			var userIdInToken = this.User.GetId();
+
+			if (userIdInToken != userId.ToString())
+			{
+				return Unauthorized();
+			}
 
 			var service = _serviceRepository.GetService(serviceId, userId);
 			if (service == null)
 			{
 				return NotFound();
+			}
+
+			// validate user is owner
+			if (service.UserId != userId)
+			{
+				return Unauthorized();
 			}
 
 			// map ServiceUpdateDto to Service
@@ -125,6 +163,13 @@ namespace Services.Controllers
 			_serviceRepository.UpdateService(service);
 			// save to db
 			_serviceRepository.SaveChanges();
+
+			// Publish ServiceUpdated event
+			await _publishEndPoint.Publish<ServiceUpdated>(new
+			{
+				Id = service.Id,
+				Info = service.Info
+			});
 
 			return NoContent();
 		}
@@ -136,18 +181,36 @@ namespace Services.Controllers
 		/// <param name="serviceId"></param>
 		/// <returns></returns>
 		[HttpDelete("{serviceId}")]
-		public ActionResult DeleteService(int userId, int serviceId)
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+		public async Task<ActionResult> DeleteService(int userId, int serviceId)
 		{
-			// TODO: validate user is logged in (JWT)
-			// TODO: validate user is owner (if not unauthorized)
+			// check if userId is the same as the current user id
+			var userIdInToken = this.User.GetId();
+
+			if (userIdInToken != userId.ToString())
+			{
+				return Unauthorized();
+			}
 
 			// get service
 			var service = _serviceRepository.GetService(serviceId, userId);
+
+			// validate user is owner
+			if (service.UserId != userId)
+			{
+				return Unauthorized();
+			}
 
 			_serviceRepository.DeleteService(service);
 
 			if (_serviceRepository.SaveChanges())
 			{
+				// Publish ServiceDeleted event
+				await _publishEndPoint.Publish<ServiceDeleted>(new
+				{
+					Id = service.Id
+				});
+
 				return Ok();
 			}
 
@@ -155,14 +218,29 @@ namespace Services.Controllers
 		}
 
 		[HttpPost("{serviceId}/photos")]
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
 		public async Task<ActionResult<PhotoDto>> AddPhoto(int userId, int serviceId, IFormFile file)
 		{
+			// check if userId is the same as the current user id
+			var userIdInToken = this.User.GetId();
+
+			if (userIdInToken != userId.ToString())
+			{
+				return Unauthorized();
+			}
+      
 			// get service object with the photos
 			var service = _serviceRepository.GetService(serviceId, userId);
 
 			if (service == null)
 			{
 				return NotFound("Service does not exist");
+			}
+
+			// validate user is owner
+			if (service.UserId != userId)
+			{
+				return Unauthorized();
 			}
 
 			// add new photo to Cloudinary
@@ -200,14 +278,29 @@ namespace Services.Controllers
 		/// <param name="photoId">photo id</param>
 		/// <returns></returns>
 		[HttpDelete("{serviceId}/photos/{photoId}")]
+		[Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
 		public async Task<ActionResult> DeletePhoto(int userId, int serviceId, int photoId)
 		{
+			// check if userId is the same as the current user id
+			var userIdInToken = this.User.GetId();
+
+			if (userIdInToken != userId.ToString())
+			{
+				return Unauthorized();
+			}
+
 			// get service object with the photos
 			var service = _serviceRepository.GetService(serviceId, userId);
 
 			if (service == null)
 			{
 				return NotFound("Service does not exist");
+			}
+
+			// validate user is owner
+			if (service.UserId != userId)
+			{
+				return Unauthorized();
 			}
 
 			// find photo
